requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.11.1
aiohttp-retry==2.9.1
aiosignal==1.3.1
attrs==24.2.0
blinker==1.9.0
boto3==1.35.68
botocore==1.35.68
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
Flask==3.1.0
frozenlist==1.5.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
MarkupSafe==3.0.2
multidict==6.1.0
propcache==0.2.0
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
s3transfer==0.10.4
six==1.16.0
twilio==9.3.6
urllib3==2.2.3
Werkzeug==3.1.3
yarl==1.17.1
